+++ src/main/kotlin/blackjack/domain/card/CardShape.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.card
 
 enum class CardShape {
     SPADE,
