+++ src/main/kotlin/blackjack/domain/card/CardShuffle.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain.card
 
 interface CardShuffle {
     fun getCards(): List<Card>
