+++ src/main/kotlin/blackjack/domain/card/Score.kt
@@ -1,4 +1,6 @@
-package blackjack
+package blackjack.domain.card
+
+import blackjack.domain.card.PlayerDeck
 
 class Score(private val playerDeck: PlayerDeck) {
