+++ src/main/kotlin/blackjack/domain/player/PlayerVo.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.domain.player
 
 class PlayerVo(val name: String, val cards: String)
