+++ src/test/kotlin/blackjack/CardTest.kt
@@ -1,5 +1,8 @@
 package blackjack
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardShape
+import blackjack.domain.card.CardSymbol
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
