+++ src/test/kotlin/blackjack/ScoreTest.kt
@@ -1,5 +1,9 @@
 package blackjack
 
+import blackjack.domain.card.Card
+import blackjack.domain.card.CardShape
+import blackjack.domain.card.CardSymbol
+import blackjack.domain.card.Score
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
